(maint) Add shared example
When running in CI, especially for containers that are very quick to
start up, it can be helpful to manage container lifecycle directly in
the example. This sets up a shared example that starts a container built
from the current directory, runs a command, checks the exit status, and
destroys the container.

The shared context should still work for many cases, but if you see
failures along the lines of 'container is not running' the shared
example can help you work through that.